fix: #30 회원가입 페이지 라우팅 수정
