+++ src/app/signup/components/ProfileImageInput.tsx
@@ -7,7 +7,7 @@ interface ProfileImageInputProps {
   onChange: (img: string | null) => void;
 }
 
-export default function ProfileImageInput({ value, onChange }: ProfileImageInputProps) {
+const ProfileImageInput: React.FC<ProfileImageInputProps> = ({ value, onChange }) => {
   const fileInputRef = useRef<HTMLInputElement>(null);
 
   const handleImgChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -24,7 +24,7 @@ export default function ProfileImageInput({ value, onChange }: ProfileImageInput
   return (
     <div className="mb-6 flex flex-col items-center">
       <div
-        className="relative mb-2 h-20 w-20 cursor-pointer rounded-full border-2 border-lavender-300 bg-gray-100 transition-all hover:ring-2 hover:ring-lavender-400 dark:border-lavender-700 dark:bg-gray-800"
+        className="relative mb-2 h-20 w-20 cursor-pointer rounded-full border-2 border-violet-300 bg-gray-100 transition-all hover:ring-2 hover:ring-violet-400 dark:border-violet-700 dark:bg-gray-800"
         onClick={handleImgClick}
         tabIndex={0}
         role="button"
@@ -32,7 +32,7 @@ export default function ProfileImageInput({ value, onChange }: ProfileImageInput
         {value ? (
           <Image src={value} alt="프로필 사진" fill className="object-cover" />
         ) : (
-          <User className="absolute left-1/2 top-1/2 h-10 w-10 -translate-x-1/2 -translate-y-1/2 text-lavender-300 dark:text-lavender-700" />
+          <User className="absolute left-1/2 top-1/2 h-10 w-10 -translate-x-1/2 -translate-y-1/2 text-violet-300 dark:text-violet-700" />
         )}
         <input
           type="file"
@@ -41,9 +41,8 @@ export default function ProfileImageInput({ value, onChange }: ProfileImageInput
           className="hidden"
           onChange={handleImgChange}
         />
-        {/* 이미지 아이콘을 input(프로필 원) 우측 하단에 완전히 보이도록 */}
         <span
-          className="absolute bottom-1 right-1 z-20 translate-x-1/4 translate-y-1/4 rounded-full bg-lavender-400 p-1.5 text-white shadow-md dark:bg-lavender-700"
+          className="absolute bottom-1 right-1 z-20 translate-x-1/4 translate-y-1/4 rounded-full bg-violet-400 p-1.5 text-white shadow-md dark:bg-violet-700"
           style={{ pointerEvents: 'none' }}
         >
           <ImageIcon className="h-4 w-4" />
@@ -52,4 +51,6 @@ export default function ProfileImageInput({ value, onChange }: ProfileImageInput
       <span className="text-xs text-gray-400 dark:text-gray-500">프로필 사진을 선택하세요</span>
     </div>
   );
-}
+};
+
+export default ProfileImageInput;
